[Game] Property move WorldManager to WorldInstance
- Moved properties that need to be instance dependent from `WorldManager` to `WorldInstance`
- WIP does not compile
